Revert the end-of-session mirror writes; the CCCD change is the fix
The reporter's controlled A/B settles it. 2.7.8-beta.15 carried the CCCD
change with the mirror writes switched off, and on a BP5465 over local BlueZ
with every proxy and the phone powered down, one fresh pairing was followed by
four connections that all resumed the retained bond: one SMP pairing total,
five successful AES-CCM encryption changes, zero CCCD disables, zero mirror
write opcodes, zero PIN or Key Missing. The measurement-triggered reconnect
decoded the new reading and the entities updated.

So the mirror writes are not required for bond retention, and they were the
riskier half: the only code in the integration that writes to a monitor's
EEPROM, with the index marker resting on one sample, on a profile shared by
HEM-7381T1-AZ, HEM-7382T1 and HEM-7388T1 whose addresses were never captured.
Nothing proven is left to justify keeping them.

Also drops the Data Pending rationale that went with them: beta.13 performed
both writes and the cuff still advertised forced_transfer afterwards, so the
pending flag is unrelated and belongs to its own issue.

The tests split accordingly. What remains covers the CCCD behaviour and moves
to a file named for it. One of those tests was already failing on the branch:
extending keep_notify_subscriptions across WLD3.0 and WLD4.0 pulled HEM-7376T1
inside the families while the test still expected it to disable on close. It
now uses a profile that really is outside.

Refs #91
